refactor(api): move and rename update-from-pix-admin controller
in team
