+++ api/lib/application/certification-center-memberships/certification-center-membership-controller.js
@@ -1,4 +1,4 @@
-import { BadRequestError, ForbiddenError } from '../../../src/shared/application/http-errors.js';
+import { ForbiddenError } from '../../../src/shared/application/http-errors.js';
 import * as certificationCenterMembershipSerializer from '../../../src/shared/infrastructure/serializers/jsonapi/certification-center-membership.serializer.js';
 import * as requestResponseUtils from '../../../src/shared/infrastructure/utils/request-response-utils.js';
 import { certificationCenterMembershipRepository } from '../../../src/team/infrastructure/repositories/certification-center-membership.repository.js';
@@ -24,32 +24,6 @@ const disableFromPixCertif = async function (request, h, dependencies = { reques
     updatedByUserId: currentUserId,
   });
   return h.response().code(204);
-};
-
-const updateFromPixAdmin = async function (
-  request,
-  h,
-  dependencies = { requestResponseUtils, certificationCenterMembershipSerializer },
-) {
-  const certificationCenterMembershipId = request.params.id;
-  const certificationCenterMembership = dependencies.certificationCenterMembershipSerializer.deserialize(
-    request.payload,
-  );
-  const pixAgentUserId = dependencies.requestResponseUtils.extractUserIdFromRequest(request);
-
-  if (certificationCenterMembershipId !== certificationCenterMembership.id) {
-    throw new BadRequestError();
-  }
-
-  const updatedCertificationCenterMembership = await usecases.updateCertificationCenterMembership({
-    certificationCenterMembershipId,
-    role: certificationCenterMembership.role,
-    updatedByUserId: pixAgentUserId,
-  });
-
-  return h.response(
-    dependencies.certificationCenterMembershipSerializer.serializeForAdmin(updatedCertificationCenterMembership),
-  );
 };
 
 const updateFromPixCertif = async function (
@@ -85,7 +59,6 @@ const updateFromPixCertif = async function (
 const certificationCenterMembershipController = {
   disableFromPixAdmin,
   disableFromPixCertif,
-  updateFromPixAdmin,
   updateFromPixCertif,
 };
