+++ api/src/team/application/certification-center-membership/certification-center-membership.admin.controller.js
@@ -1,0 +1,36 @@
+import { usecases } from '../../../../lib/domain/usecases/index.js';
+import { BadRequestError } from '../../../shared/application/http-errors.js';
+import * as certificationCenterMembershipSerializer from '../../../shared/infrastructure/serializers/jsonapi/certification-center-membership.serializer.js';
+import { requestResponseUtils } from '../../../shared/infrastructure/utils/request-response-utils.js';
+
+const updateRole = async function (
+  request,
+  h,
+  dependencies = { requestResponseUtils, certificationCenterMembershipSerializer },
+) {
+  const certificationCenterMembershipId = request.params.id;
+  const certificationCenterMembership = dependencies.certificationCenterMembershipSerializer.deserialize(
+    request.payload,
+  );
+  const pixAgentUserId = dependencies.requestResponseUtils.extractUserIdFromRequest(request);
+
+  if (certificationCenterMembershipId !== certificationCenterMembership.id) {
+    throw new BadRequestError();
+  }
+
+  const updatedCertificationCenterMembership = await usecases.updateCertificationCenterMembership({
+    certificationCenterMembershipId,
+    role: certificationCenterMembership.role,
+    updatedByUserId: pixAgentUserId,
+  });
+
+  return h.response(
+    dependencies.certificationCenterMembershipSerializer.serializeForAdmin(updatedCertificationCenterMembership),
+  );
+};
+
+const certificationCenterMembershipAdminController = {
+  updateRole,
+};
+
+export { certificationCenterMembershipAdminController };
