+++ api/src/team/application/certification-center-membership/certification-center-membership.admin.route.js
@@ -1,12 +1,12 @@
-import { certificationCenterMembershipController } from '../../../../lib/application/certification-center-memberships/certification-center-membership-controller.js';
 import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
+import { certificationCenterMembershipAdminController } from './certification-center-membership.admin.controller.js';
 
 export const certificationCenterMembershipAdminRoutes = [
   {
     method: 'PATCH',
     path: '/api/admin/certification-center-memberships/{id}',
     config: {
-      handler: (request, h) => certificationCenterMembershipController.updateFromPixAdmin(request, h),
+      handler: (request, h) => certificationCenterMembershipAdminController.updateRole(request, h),
       pre: [
         {
           method: (request, h) =>
